requirements.txt
# pymysql==1.0.2
pandas==1.1.4
sqlalchemy==1.4.29
confluent-kafka==1.7.0
faker==9.9.0
requests==2.26.0
jsonschema==4.4.0